# fix Dict `sizehint!` to allow enough space for the requested length