fixes #9909
noticed in #37208

Test code:
```
function grow_dict(szh::Bool)
    d = Dict{Int,Nothing}()
    szh && sizehint!(d, 32768)
    for i in 1:32768
        d[i] = nothing
    end
    length(d.keys)
end
```
before:
```
julia> @btime grow_dict(true)
  704.747 μs (12 allocations: 1.41 MiB)
131072

julia> @btime grow_dict(false)
  702.613 μs (27 allocations: 770.16 KiB)
65536
```
after:
```
julia> @btime grow_dict(true)
  464.426 μs (7 allocations: 576.66 KiB)
65536

julia> @btime grow_dict(false)
  702.545 μs (27 allocations: 770.16 KiB)
65536
```

@bkamins You'll probably want to rerun the benchmarks for `allunique`!